Require at least size 1 for cstr destination
Calling ss8_copy_to_cstr() with bufsize == 0 is likely an indication of
a bug (the result will not be a valid C string), so add an assertion
against it. But do the "nice" thing (do nothing) when assertions are
disabled.

If this turns out to be too strict for convenience, we can always relax
the precondition later while being compatible; let's try to be strict
for now.