Flush unit test output after each test
Surprisingly this is not enabled by default in Unity, so if a test
crashes, the output can be lost from some previous tests as well.

Ideally we would set this as the subproject's default options, but this
doesn't seem to work with the Unity wrap. So set it as a global default
option instead. It will get lost if c_args is globally overridden, but
the important thing is that flushing occurs when run in CI.